Add new method to find the plugin.
It will read the first line of ~/.ember/webeber.path,
which should be a path to the ember prefix.
NOTE: in browser ~ it not always the users home,
so I use HOME environment variable.
This is used on releases, until I make webember.conf file.
It can be overwritten with WEBEMBER_PREFIX environment variable.